requirements.txt
Package                       Version
----------------------------- -------
backcall                      0.1.0  
coverage                      4.5.3  
decorator                     4.4.0  
Django                        1.11.6 
django-extensions             2.1.7  
django-shell-plus             1.1.7  
django-tastypie               0.14.2 
djangorestframework           3.9.4  
djangorestframework-simplejwt 4.3.0  
ipython                       7.5.0  
ipython-genutils              0.2.0  
jedi                          0.13.3 
parso                         0.4.0  
pexpect                       4.7.0  
pickleshare                   0.7.5  
pip                           19.1.1 
prompt-toolkit                2.0.9  
ptyprocess                    0.6.0  
Pygments                      2.4.0  
PyJWT                         1.7.1  
python-dateutil               2.8.0  
python-mimeparse              1.6.0  
pytz                          2019.1 
setuptools                    41.0.1 
six                           1.12.0 
traitlets                     4.3.2  
wcwidth                       0.1.7  
wheel                         0.33.4 
